feat(groupchat,core,cli): add group chat with members, multi-peer reliability, and UX tweaks
- Core: ACK-to-origin, per-peer Sender map, send_bytes_to(dst) API, multi-tick
- Net bind: wire emit_pdu_to -> eth_send_pdu_to
- Chat (1:1): dynamic target by alias, periodic silent HELLO, timestamps, alias display
- Groupchat: new interactive mode, default send to all online, /members list|add|del|clear, /online, /contacts, /sendfile
- CLI: remove deprecated /all and /allfile; drop /discover inside chat; config no longer asks for dst MAC
- Discover: use one-shot HELLO PDU (fire-and-forget) for broadcast
- Reassembly: fix 256-chunk cap (uint8_t -> uint32_t in byte count loop)
- README: update docs with groupchat and script usage
- Tooling: script.sh with up [N], attach, down, clean, rebuild; remove docker-compose.yml

+++ src/reassembly.cpp
@@ -147,7 +147,7 @@ namespace linkchat
             return false;
         
         size_t total_bytes = 0;
-        for (uint8_t chunk = 0; chunk < msgs_[msg_id].total; chunk++)
+        for (uint32_t chunk = 0; chunk < msgs_[msg_id].total; chunk++)
         {
             total_bytes += msgs_[msg_id].chunks[chunk].size();
         }
